Allow color value passed to svg generator
Added code to allow for a color value to be passed to genomicDataRectangleGenerator.
If the 'percent' input is an object with keys 'value' and 'colorValue', the
text will be the value, and the color of the svg will be the colorValue.
Otherwise the text will be the percent and the existing red/blue color
map is used.

Also moved the initialization of the NGCHM so that the editor is defined
before initialization and fixed the getLabels so the editor is defined.