Pass highlight info to NGCHM & Bug fix
Added code for highlighting back and forth between the NGCHM and the
Pathway: the menu items 'Highlight -> Highlight Selected' and
'Highlight -> Remove All Highlights' now also affect the highlighting
on the NGCHM. In addition, clicking labels on the heatmap will highlight
corresponding labels in Pathway.

Part of accomplishing that: moved the NGCHM initialization into PathwayActions,
because it is needed there but not in react-pathway-mapper. Also added
an editorHandler for the NGCHM (in the model of the editor handler for
PathwayActions and react-pathway-mapper).

Bug fix: declaration of 'text' variable in GenomicDataOverlayManager.ts