Give toast messages when adding NGCHM datasets
Using the existing toast installation to give a message when
a dataset is loaded from the NG-CHM, and give a warning if the
dataset name is being reused (i.e. the dataset has been
overwritten).

Added back the toast css that was removed in 12bdf42. Removal of
that css caused the toast messages that appear, for example,
when a dataset is loaded from cBioPortal, to show up in the
lower left of the page, and not disapear automatically.