Fix bug: checking if dataset already in profiles
Fix a bug for the list of knownDatasets when trying decide if
the datasetLabel should be added to the profiles.
The fix: get list of existing profileIds directly from profiles
(instead of pretenting to keep track of them in NGCHM.tsx).

Changed variable name 'knownDatasets' to 'existingProfiles' in an
attempt to be more descriptive.

Closes #2