Only use type as 'GENE' in loading NDEx pathway
NDEx pathways allow for any string as a node 'type'. However
editor.getGeneSymbols() will only return items with type == 'GENE'.
Since we use the getGeneSymbols() function to get nodes for the purpose
of the NGCHM interaction, we specify in the loader for NDEx that all the
node types are 'GENE' in LoadFromExternalDatabase.

This commit also makes some cosmetic changes:

- node hight and width changed to those of PathwayMapper examples so
  that loaded genomic data looks better
- output file name changed to be that of name of loaded pathway
  (in LoadFromExternalDatabase)