Add NDEx pathways to Network menu
This is for the situation in which an NGCHM has NDEx pathway information
in the form of UUIDs. Upon receiving this information,
NGCHM.ndexSummary() is called to get the name of the pathway from NDEx.
That name is then displayed in the menu under
Network -> External Databse -> NDEx.

The pathway reference information from an NGCHM is a mobx observable
(so that the Menubar component is re-rendered when the values change)

In doing this, moved NGCHM from a prop to PathwayActions to instead be
a prop of the PathwayMapper component (i.e. in react-pathway-mapper.tsx)
and Menubar component.

The functionality of LoadFromExternal was moved to NGCHM.tsx in order to
minimize our footprint in the code (by needing one less functional
class and one less file). The NDEx modal was removed (for the moment
at least)