Add CSS for disabled MenuItem
Added CSS for disabled MenuItem in pmv2.css. This was done because there
can be NDEx pathways that are not loadable (because too many nodes,
or invalid UUID), and we want to denote that in the dropdown (here in
Network -> NDEx -> <pathway name>).

Removed the separate file for ngchm.css because it seemed easier
to add this single CSS entry in the now-unminified pmv2.css.

Removed the superfluous 'disabledMenuItem' class in favor of simply
using 'disabled', which was already part of the existing CSS scheme.