Check UUID format & first loaded pathway
Added checks for proper format and sanitization of UUIDs coming from the
NGCHM.

Also added a 'loadedFirstValidPathway' property to account for
situations where the first UUID from the NGCHM is not valid. In this
type of situation the desired behaviour is to load the first valid
pathway. (All pathways are still listed in the Network -> NDEx menu,
with appropriate notes about how/if the pathway was invalid.)